fastapi-ml/main_new.py
# main_new.py
# This file will contain the new implementation of the FastAPI application,
# aligned with the wide-format database schema and multi-model strategy.

# We will build this file phase by phase.

from fastapi import FastAPI, HTTPException, Depends
from pydantic import BaseModel
from typing import Dict, Any, List, Optional
import pandas as pd
import numpy as np
import xgboost as xgb
import joblib
import psycopg2
from psycopg2.extras import RealDictCursor
from dotenv import load_dotenv
import os
from datetime import datetime

# Load environment variables
load_dotenv()

# --- Database Configuration ---
DB_HOST = os.getenv("DB_HOST", "postgres")
DB_NAME = os.getenv("POSTGRES_DB_CORE")
DB_USER = os.getenv("POSTGRES_USER")
DB_PASSWORD = os.getenv("POSTGRES_PASSWORD")

def get_db_connection():
    """Establishes a connection to the PostgreSQL database."""
    try:
        conn = psycopg2.connect(
            host=DB_HOST,
            database=DB_NAME,
            user=DB_USER,
            password=DB_PASSWORD,
            cursor_factory=RealDictCursor
        )
        return conn
    except Exception as e:
        print(f"Database connection error: {e}")
        raise HTTPException(status_code=500, detail="Database connection error.")

# --- FastAPI App Initialization ---
app = FastAPI(
    title="Energy AI Platform - Core API v2",
    version="2.0.0",
    description="API for training and serving energy consumption models with a multi-model strategy."
)

# --- Health Check Endpoints ---
@app.get("/", tags=["Health"])
def read_root():
    """Root endpoint providing a welcome message."""
    return {"message": "Welcome to the Energy AI Platform Core API v2!"}

@app.get("/health", tags=["Health"])
def health():
    """Health check endpoint to verify service status."""
    return {"status": "ok", "message": "API is healthy"}

# --- Development will continue here, phase by phase ---
