fastapi-ml/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pandas
scikit-learn
python-dotenv
xgboost
requests